// web/src/api/generated/models/class-schedule-item-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ClassScheduleItemResponse {
    'id'?: string;
    'settingId'?: string;
    'classId'?: string;
    'userId'?: string;
    'courseType'?: number;
    'courseTypeDesc'?: string;
    'courseName'?: string;
    'teacherName'?: string;
    'location'?: string;
    'courseId'?: string;
    'teacherId'?: string;
    'dayOfWeek'?: number;
    'startSection'?: number;
    'endSection'?: number;
    'startWeek'?: number;
    'endWeek'?: number;
    'weekType'?: number;
    'weekTypeDesc'?: string;
    'color'?: string;
    'remark'?: string;
    'createTime'?: string;
    'updateTime'?: string;
}